+++ include/phoxi_camera/PhoXiInterface.h
@@ -1,0 +1,17 @@
+//
+// Created by controller on 1/11/18.
+//
+
+#ifndef PROJECT_PHOXIINTERFACE_H
+#define PROJECT_PHOXIINTERFACE_H
+
+
+class PhoXiInterface {
+
+private:
+
+
+};
+
+
+#endif //PROJECT_PHOXIINTERFACE_H